+++ scm-ui/ui-components/src/repos/index.ts
@@ -52,6 +52,7 @@ export { default as RepositoryAvatar } from "./RepositoryAvatar";
 export { default as RepositoryEntry } from "./RepositoryEntry";
 export { default as RepositoryEntryLink } from "./RepositoryEntryLink";
 export { default as JumpToFileButton } from "./JumpToFileButton";
+export { default as CommitAuthor } from "./CommitAuthor";
 
 export {
   File,